migrations, public shipment handler updates, test data gen status change
